+++ geoh5py/shared/utils.py
@@ -1333,108 +1333,6 @@ def validate_normalized_vector(value: np.ndarray) -> np.ndarray:
     return value
 
 
-def all_subclasses(type_object: T) -> list[T]:
-    """
-    Recursively find all subclasses of input type object.
-
-    :param type_object: Class containing possibly nested subclasses.
-
-    :returns: List of all the subclasses for the entire hierarchy.
-    """
-    collection = []
-    subclasses = type_object.__subclasses__()
-    collection += subclasses
-    for subclass in subclasses:
-        collection += all_subclasses(subclass)
-    return collection
-
-
-def model_fields_difference(
-    parent: T,
-    child: T,
-) -> set[str]:
-    """
-    Find fields added in the child class.
-
-    :param parent: Parent class.
-    :param child: Child class.
-
-    :returns: All fields in the child class but not the parent.
-    """
-    return set(child.model_fields) - set(parent.model_fields)
-
-
-def indicator_attributes(
-    parent: T, children: list[T]
-) -> tuple[list[set[str]], list[set[str]]]:
-    """
-    List all the strong (required) and weak (optional) indicator attributes.
-
-    :param parent: Parent class.
-    :param children: A children of the parent.
-
-    :return strong_indicators: List of all required attributes not in the
-        parent for each of the children.
-    :return weak_indicators: List of all optional attributes not in the
-        parent for each of the children.
-    """
-    strong_indicators = []
-    weak_indicators = []
-    for child in children:
-        indicators = model_fields_difference(parent, child)
-        strong_indicators_set = set()
-        weak_indicators_set = set()
-        for field in indicators:
-            if child.model_fields[field].is_required():
-                strong_indicators_set.add(field)
-            else:
-                weak_indicators_set.add(field)
-        strong_indicators.append(strong_indicators_set)
-        weak_indicators.append(weak_indicators_set)
-
-    return strong_indicators, weak_indicators
-
-
-def best_match_subclass(parent: T, children: list[T], data: dict[str, Any]) -> T:
-    """
-    Choose the subclass with the best matching attributes.
-
-    :param parent: Parent class.
-    :param children: All children of the parent.
-    :param data: Dictionary of attribute data to match.
-
-    :returns: The best matching subclass.
-    """
-
-    strong_indicators, weak_indicators = indicator_attributes(parent, children)
-    n_strong = np.array(
-        [len(strong_indicators[i].intersection(data)) for i, c in enumerate(children)]
-    )
-    n_weak = np.array(
-        [len(weak_indicators[i].intersection(data)) for i, c in enumerate(children)]
-    )
-
-    total_indicators = n_strong + n_weak
-    candidates = np.array(children)[total_indicators == np.max(total_indicators)]
-    if len(candidates) == 1:
-        return candidates[0]
-
-    candidates = np.array(children)[n_strong == np.max(n_strong)]
-    if len(candidates) == 1:
-        return candidates
-    if len(candidates) > 1 and len(candidates) < len(children):
-        return candidates[np.argmin([len(k.model_fields) for k in candidates])]
-
-    candidates = [c for i, c in enumerate(children) if not strong_indicators[i]]
-    value = data.get("value", None)
-    for candidate in candidates:
-        annotation = candidate.model_fields["value"].annotation
-        if type(value) == annotation:  # pylint: disable=unidiomatic-typecheck
-            return candidate
-
-    raise ValueError(f"Could not match data: {data} to any of the children.")
-
-
 def are_coplanar(points: np.ndarray, tol: float = 1e-6) -> bool:
     """
     Check if a set of points are coplanar.
