+++ geoh5py/ui_json/__init__.py
@@ -26,3 +26,5 @@ from __future__ import annotations
 from .input_file import InputFile
 from .utils import monitored_directory_copy
 from .validation import InputValidation
+from .forms import BaseForm
+from .ui_json import BaseUIJson
