+++ tests/ui_json/forms_test.py
@@ -24,7 +24,7 @@ import uuid
 
 import numpy as np
 import pytest
-from pydantic import ValidationError
+from pydantic import BaseModel, ValidationError
 
 from geoh5py import Workspace
 from geoh5py.groups import GroupTypeEnum, PropertyGroup
@@ -48,6 +48,9 @@ from geoh5py.ui_json.forms import (
     ObjectForm,
     RadioLabelForm,
     StringForm,
+    all_subclasses,
+    best_match_subclass,
+    model_fields_difference,
 )
 from geoh5py.ui_json.ui_json import BaseUIJson
 
@@ -706,3 +709,71 @@ def test_something():
             "verbose": 2,
         }
     )
+
+
+def test_all_subclasses():
+    class TestOne:
+        pass
+
+    class TestTwo(TestOne):
+        pass
+
+    class TestThree(TestTwo):
+        pass
+
+    class TestFour(TestOne):
+        pass
+
+    assert len(all_subclasses(TestOne)) == 3
+    assert all(k in all_subclasses(TestOne) for k in [TestTwo, TestThree, TestFour])
+    assert len(all_subclasses(TestTwo)) == 1
+    assert all_subclasses(TestTwo)[0] == TestThree
+    assert all_subclasses(TestThree) == []
+    assert all_subclasses(TestFour) == []
+
+
+def test_model_fields_difference():
+    class MyParent(BaseModel):
+        a: str
+
+    class MyChild(MyParent):
+        b: int
+        c: float
+
+    class MyOtherChild(MyChild):
+        d: float
+        e: float = 1.0
+
+    diff = model_fields_difference(MyParent, MyChild)
+    assert diff == {"b", "c"}
+    diff = model_fields_difference(MyParent, MyOtherChild)
+    assert diff == {"b", "c", "d", "e"}
+
+
+def test_best_match_subclass():
+    class Parent(BaseModel):
+        a: str
+        b: int
+
+    class ChildA(Parent):
+        c: float
+
+    class ChildB(ChildA):
+        d: float
+
+    class ChildC(Parent):
+        e: float
+
+    best_match = best_match_subclass(
+        parent=Parent,
+        data={"a": "test", "b": 1, "c": 1.0, "d": 1.0},
+    )
+
+    assert best_match == ChildB
+
+    best_match = best_match_subclass(
+        parent=Parent,
+        data={"a": "test", "b": 1, "c": 1.0},
+    )
+
+    assert best_match == ChildA
